Add proof of concept file navigation and selection
The script "libresmart.sh" runs a menu loop showing the contents of the currently selected directory. If a file is selected, the default editor is used to open it.
If this is copied to the default user's folder on a fresh Raspberry Pi OS install, and the .bashrc is modified according to the comments in the file, this will turn the device into a dedicated writing machine.